Add variables: onlynames will be set to 1 if '-l' option is given, else 0; fmatch will count the number of matches found in the current file.
Add non-parameter option 'l' to string "shortopts."

Set "onlynames" to 1 when option switch 'l' is recognized.

Re-set fmatch to 0 after each call to fopen; not needed in the stdin case
because stdin is never re-opened, and fmatch is initialized to 0 by its
declaration.

Increment fmatch whenever nmatch is incremented (in two places).  Change
match-printing logic (in two places) to print only the filename (or
"(standard input)" if no filename) when "onlynames" is set; and to print
"(standard input)" when no filename is given but filename labels are
explicitly requested (compatibility with potential '-H' option support).
After printing the match, optimize speed for "onlynames" cases (in two
places) by not reading the input file once a match is found (state S_NEXTF).

In function scan_block, implement the S_NEXTF state by setting pointer p to
the end-of-buffer pointer plim.  For scan_read cases, the while loop in
function scan_read has been modified to terminate when it detects the S_NEXTF
state.